+++ xwiki-platform-core/xwiki-platform-security/xwiki-platform-security-authorization/xwiki-platform-security-authorization-bridge/src/test/java/org/xwiki/security/authorization/testwikibuilding/LegacyTestWiki.java
@@ -1073,6 +1073,8 @@ public class LegacyTestWiki extends AbstractTestWiki
                     allowing(mockedDocument)
                         .getXObjects(XWikiGroupsDocumentInitializer.XWIKI_GROUPS_DOCUMENT_REFERENCE);
                     will(returnValue(Collections.emptyList()));
+                    allowing(mockedDocument).isRequiredRightsActivated();
+                    will(returnValue(false));
                 }
             });
